developer/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.3
numpy==1.22.0
opencv-python==4.6.0.66
mysql-connector-python==8.0.30
Flask-Session